+++ src/main/java/org/openlmis/stockmanagement/web/SearchParams.java
@@ -1,0 +1,68 @@
+/*
+ * This program is part of the OpenLMIS logistics management information system platform software.
+ * Copyright © 2017 VillageReach
+ *
+ * This program is free software: you can redistribute it and/or modify it under the terms
+ * of the GNU Affero General Public License as published by the Free Software Foundation, either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Affero General Public License for more details. You should have received a copy of
+ * the GNU Affero General Public License along with this program. If not, see
+ * http://www.gnu.org/licenses.  For additional information contact info@OpenLMIS.org.
+ */
+
+package org.openlmis.stockmanagement.web;
+
+import java.util.Collection;
+import lombok.NoArgsConstructor;
+import org.apache.commons.collections4.MapUtils;
+import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.util.MultiValueMap;
+
+@NoArgsConstructor
+public class SearchParams {
+
+  public static final String PAGE = "page";
+  public static final String SIZE = "size";
+  public static final String SORT = "sort";
+  public static final String ACCESS_TOKEN = "access_token";
+
+  private MultiValueMap<String, Object> params;
+
+  /**
+   * Constructs new SearchParams object from {@code MultiValueMap}.
+   */
+  public SearchParams(MultiValueMap<String, Object> queryMap) {
+    if (queryMap != null) {
+      params = new LinkedMultiValueMap<>(queryMap);
+      params.remove(PAGE);
+      params.remove(SIZE);
+      params.remove(SORT);
+      params.remove(ACCESS_TOKEN);
+    } else {
+      params = new LinkedMultiValueMap<>();
+    }
+  }
+
+  public boolean containsKey(String key) {
+    return params.containsKey(key);
+  }
+
+  public Object getFirst(String key) {
+    return params.getFirst(key);
+  }
+
+  public LinkedMultiValueMap<String, Object> asMultiValueMap() {
+    return new LinkedMultiValueMap<>(params);
+  }
+
+  public Collection<String> keySet() {
+    return params.keySet();
+  }
+
+  public boolean isEmpty() {
+    return MapUtils.isEmpty(params);
+  }
+}